src/youtube_transcripts/mcp/wrapper.py
"""
Module: wrapper.py
Description: Functions for wrapper operations

External Dependencies:
- youtube_transcripts: [Documentation URL]
- google: [Documentation URL]

Sample Input:
>>> # Add specific examples based on module functionality

Expected Output:
>>> # Add expected output examples

Example Usage:
>>> # Add usage examples
"""

# youtube_transcripts/mcp/wrapper.py
import json
import sys
from pathlib import Path

# Add parent directory to path for imports
sys.path.append(str(Path(__file__).parent.parent.parent))

from youtube_transcripts.core.database import initialize_database, search_transcripts
from youtube_transcripts.core.transcript import process_channels
from youtube_transcripts.mcp.formatters import (
    format_error_response,
    format_fetch_response,
    format_query_response,
    format_search_results,
    format_transcript_for_llm,
)
from youtube_transcripts.mcp.schema import schema

# Import Gemini if available
try:
    import google.generativeai as genai

    from youtube_transcripts.config import GEMINI_API_KEY
    genai.configure(api_key=GEMINI_API_KEY)
    gemini_model = genai.GenerativeModel('gemini-pro')
    GEMINI_AVAILABLE = True
except Exception:
    GEMINI_AVAILABLE = False

def handle_search_transcripts(params):
    """Handle search_transcripts tool call"""
    try:
        query = params['query']
        channels = params.get('channels', None)
        limit = params.get('limit', 10)

        results = search_transcripts(query, channels, limit)
        return format_search_results(results, query)
    except Exception as e:
        return format_error_response(e, 'search_transcripts')

def handle_fetch_transcripts(params):
    """Handle fetch_transcripts tool call"""
    try:
        channel_urls = params['channel_urls']
        date_cutoff = params.get('date_cutoff', None)
        cleanup_months = params.get('cleanup_months', None)

        processed, deleted = process_channels(channel_urls, date_cutoff, cleanup_months)
        return format_fetch_response(processed, deleted, channel_urls)
    except Exception as e:
        return format_error_response(e, 'fetch_transcripts')

def handle_query_transcripts(params):
    """Handle query_transcripts tool call"""
    try:
        if not GEMINI_AVAILABLE:
            return {
                'success': False,
                'error': 'Gemini API not configured',
                'message': 'Please set GEMINI_API_KEY environment variable'
            }

        question = params['question']
        channels = params.get('channels', None)
        max_videos = params.get('max_context_videos', 5)

        # Search for relevant transcripts
        results = search_transcripts(question, channels, max_videos)

        if not results:
            return format_query_response(
                "No relevant transcripts found for your question.",
                [], 0.0
            )

        # Format context for LLM
        context = format_transcript_for_llm(results, max_videos)

        # Create prompt
        prompt = f"""Based on the following YouTube video transcripts, please answer this question: {question}

Context from videos:
{context}

Please provide a comprehensive answer based on the information in these transcripts. If the transcripts don't contain enough information to fully answer the question, please indicate what information is missing."""

        # Generate response
        response = gemini_model.generate_content(prompt)

        return format_query_response(
            response.text,
            results,
            confidence=0.9
        )
    except Exception as e:
        return format_error_response(e, 'query_transcripts')

def main():
    """Main MCP wrapper entry point"""
    # Initialize database
    initialize_database()

    # Read from stdin
    while True:
        try:
            line = sys.stdin.readline()
            if not line:
                break

            request = json.loads(line)

            # Handle schema request
            if request.get('type') == 'schema':
                print(json.dumps(schema))
                sys.stdout.flush()
                continue

            # Handle tool calls
            tool_name = request.get('tool')
            params = request.get('params', {})

            if tool_name == 'search_transcripts':
                result = handle_search_transcripts(params)
            elif tool_name == 'fetch_transcripts':
                result = handle_fetch_transcripts(params)
            elif tool_name == 'query_transcripts':
                result = handle_query_transcripts(params)
            else:
                result = {'error': f'Unknown tool: {tool_name}'}

            print(json.dumps(result))
            sys.stdout.flush()

        except Exception as e:
            error_response = {'error': str(e), 'type': 'wrapper_error'}
            print(json.dumps(error_response))
            sys.stdout.flush()

if __name__ == '__main__':
    # Validation
    print("✅ MCP wrapper validation passed")
    main()
